Added license file
Since this is a public repository, we really need to add a license. I've opted for the GPL lesser 3 license. Basically it means that if anyone uses our code, or modifies the code, they have to provide the source and maintain the license. However, if they derive work from our code, they only need to include source for our (modified) code and not their derived work.

What this means if that if someone uses SFA in their own simulations, we won't know what they did. The other option is to go full GPL, which means all derived work needs to be GPL. This could technically pose problems for ourselves.

Have a think!